requirements.txt
streamlit
groq
pycountry
sentence-transformers
faiss-cpu
